requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
xgboost>=1.5.0
numpy>=1.20.0 